refactor(worker): extract the catch-up diarization stage from pipeline.py into catch_up_diarization.py
